+++ CHANGELOG.md
@@ -8,5 +8,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.0.0] - 2020-09-10
 ### Added
-- Generic Singleton
+- Generic MonoBehavior Singleton
+- Generic ScriptableObject Singleton
 - Properties to override name/instantiation/prefab location